finito registrazione esercente e cliente

+++ bookpay/book-pay-prenotazione-service/src/main/java/com/book/cloud/repository/ClienteRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.book.cloud.entity.DClienti;
 
-public interface ClienteRepository extends CrudRepository<DClienti, Long>{
+public interface ClienteRepository extends CrudRepository<DClienti, Integer>{
 
 }
